moved colpali resource to colpali folder
