+++ aidial_rag/documents.py
@@ -27,7 +27,7 @@ from aidial_rag.image_processor.extract_pages import is_image
 from aidial_rag.index_storage import IndexStorage
 from aidial_rag.print_stats import print_chunks_stats
 from aidial_rag.request_context import RequestContext
-from aidial_rag.resources.colpali_model_resource import ColpaliModelResource
+from aidial_rag.retrievers.colpali_retriever.colpali_model_resource import ColpaliModelResource
 from aidial_rag.resources.dial_limited_resources import DialLimitedResources
 from aidial_rag.retrievers.bm25_retriever import BM25Retriever
 from aidial_rag.retrievers.colpali_retriever.colpali_retriever import (
