+++ aidial_rag/retrievers/colpali_retriever/colpali_model_resource.py
