+++ tests/test_attachment_stored.py
@@ -12,7 +12,7 @@ from aidial_rag.documents import load_document
 from aidial_rag.errors import InvalidDocumentError
 from aidial_rag.index_storage import IndexStorage
 from aidial_rag.request_context import RequestContext
-from aidial_rag.resources.colpali_model_resource import ColpaliModelResource
+from aidial_rag.retrievers.colpali_retriever.colpali_model_resource import ColpaliModelResource
 from aidial_rag.resources.dial_limited_resources import DialLimitedResources
 from tests.utils.user_limits_mock import user_limits_mock
